requirements.txt
flask==3.0.3
anthropic==0.34.0
python-dotenv==1.0.1
weasyprint==62.3
razorpay==1.4.1
gunicorn==22.0.0
